+++ R/survival.R
@@ -49,14 +49,10 @@ get_probs_from_surv.flexsurvreg <- function(x, cycle,
   )
   
   type <- match.arg(type)
-  ##    if(pred_type == "surv" & use_km_until == 0)
-  ##      stop('prediction of "surv" only with Kaplan-Meier
-  ##           probabilities (use_km_until > 0)')
   
-  #get times in absolute (not Markov) units
   times_surv <- cycle_length * c(0, cycle)
   
-  preds <- rep(NA, length(cycle))
+  res <- rep(NA, length(cycle))
   
   prob_type <- ifelse(
     cycle <= km_limit,
@@ -101,9 +97,9 @@ get_probs_from_surv.flexsurvreg <- function(x, cycle,
       ## setting 0 beyond what we have because NA throws an error
       ## need to check through this more thoroughly
       KM_tr_prob = c(KM_tr_prob, rep(0, delta))
-      preds[use_km] <- KM_tr_prob[use_km]
+      res[use_km] <- KM_tr_prob[use_km]
     } else if (type == "surv") {
-      preds[use_km] <- km$surv[-length(km$surv)][use_km]
+      res[use_km] <- km$surv[-length(km$surv)][use_km]
     }
   }
   
@@ -117,65 +113,49 @@ get_probs_from_surv.flexsurvreg <- function(x, cycle,
           type = "cumhaz")
         )
         pred_tr_prob <- 1 - exp(-diff(tmp$est))
-        preds[use_pred] <- pred_tr_prob[use_pred]
+        res[use_pred] <- pred_tr_prob[use_pred]
       } else if (type == "surv") {
-        preds[use_pred] <- flexsurv::summary.flexsurvreg(
-          x,
-          t = times_surv,
+        res[use_pred] <- flexsurv::summary.flexsurvreg(
+          x, t = times_surv,
           type = "survival")[[1]]$est[-1][use_pred]
       }
     }
   }
   
   stopifnot(
-    all(! is.na(preds))
+    all(! is.na(res))
   )
   
-  preds
+  res
 }
 
 #' @rdname get_probs_from_surv
 #' @export
-get_probs_from_surv.list <- function(x, cycle,
-                                     km_limit = 0,
-                                     cycle_length = 1,
-                                     type = c("prob", "surv")) {
+get_probs_from_surv.surv_dist <- function(x, cycle,
+                                          cycle_length = 1,
+                                          type = c("prob", "surv")) {
   type <- match.arg(type)
   
   stopifnot(
-    "dist_name" %in% names(x),
     all(cycle > 0)
   )
   
-  this_fn <- paste0("r", x$dist_name)
-  
-  correct_names <- setdiff(
-    names(x), "dist_name"
-  ) %in% names(formals(this_fn))
-  
-  if (! all(correct_names)) {
-    stop(sprintf(
-      "Incorrect arguments: %s.",
-      paste(setdiff(
-        names(x), "dist_name"
-      )[! correct_names])))
-  }
-  
   times_surv <- cycle_length * c(0, cycle)
-  Hfn <- get(paste0("H", x$dist_name),
+  
+  Hf <- get(paste0("H", x$distribution),
              envir = asNamespace("flexsurv"))
-  args <- x[- match("dist_name", names(x))]
-  args <- as.list(args)
+  
+  args <- x[- match("distribution", names(x))]
   args[["x"]] <- times_surv
-  cumhaz <- do.call(Hfn, args)
+  cumhaz <- do.call(Hf, args)
   
   if(type == "prob") {
-    pred_tr_prob <- 1 - (1 / exp(diff(cumhaz)))
-  } else if (type == "surv") {
-    pred_tr_prob <- exp(-cumhaz)[-1]
+    res <- 1 - (1 / exp(diff(cumhaz)))
+  } else {
+    res <- exp(-cumhaz)[-1]
   }
   
-  pred_tr_prob
+  res
 }
 
 #' Define a Survival Distribution
@@ -191,11 +171,42 @@ get_probs_from_surv.list <- function(x, cycle,
 #'
 #' @examples
 #' 
-#' define_surv_dist(distribution = "exp", lambda = .5)
+#' define_survival(distribution = "exp", rate = .5)
+#' define_survival(distribution = "gompertz", rate = .5, shape = 1)
 #' 
-define_surv_dist <- function(distribution = c("exp", "weibull"),
-                             ...) {
+define_survival <- function(distribution = c("exp", "weibull",
+                                             "lnorm", "gamma", 
+                                             "gompertz", "gengamma"),
+                            ...) {
+  
   distribution <- match.arg(distribution)
+  
+  list_arg <- list(...)
+  
+  if (distribution %in% c("exp", "weibull",
+                          "lnorm", "gamma")) {
+    env_f <- asNamespace("stats")
+  } else {
+    if (! requireNamespace("flexsurv")) {
+      stop("'flexsurv' package required.")
+    }
+    env_f <- asNamespace("flexsurv")
+  }
+  
+  rf <- get(paste0("r", distribution),
+            envir = env_f)
+  
+  names_fun <- setdiff(names(list_arg), "distribution")
+  names_par <- setdiff(names(formals(rf)), "n")
+  
+  correct_names <- names_fun %in% names_par
+  
+  if (! all(correct_names)) {
+    stop(sprintf(
+      "Incorrect argument%s: %s.",
+      plur(sum(! correct_names)),
+      paste(names_fun[! correct_names], collapse = ", ")))
+  }
   
   structure(
     list(
@@ -204,99 +215,4 @@ define_surv_dist <- function(distribution = c("exp", "weibull"),
     ),
     class = "surv_dist"
   )
-}
-
-#' Compute Counts for Partitioned Survival Models
-#' 
-#' Compute count of individual in each state per cycle based
-#' on a partitioned survival model.
-#' 
-#' The elements of \code{x} should be called \code{pfs} and 
-#' \code{os} for progression-free survival and overall
-#' survival respectively.
-#' 
-#' See \code{\link{get_probabilities}} for further 
-#' information on \code{km_limit}, and the
-#' \code{markov_cycle} arguments.
-#' 
-#' @param x A list of \code{flexsurvreg} objects, or a list
-#'   of lists specifying distributions.
-#' @param km_limit To what time should Kaplan-Meier 
-#'   estimates be used? Model predictions will be used 
-#'   thereafter.
-#' @param num_patients The number of patients being modeled.
-#' @param state_names Name of the states to be assigned to 
-#'   the counts.
-#'   
-#' @return A \code{cycle_counts} object.
-#'   
-#' @keywords internal
-compute_counts_part_surv <- function(x, km_limit,
-                                     num_patients,
-                                     markov_cycle, markov_cycle_length,
-                                     state_names) {
-  
-  stopifnot(length(use_km_until) == 1)
-  
-  pfs_surv <- get_probabilities(
-    x$pfs, 
-    km_limit = km_limit,
-    cycle = markov_cycle,
-    cycle_length = markov_cycle_length, 
-    type = "surv"
-  )
-  
-  os_surv <- get_probabilities(
-    x$os, 
-    km_limit = km_limit,
-    cycle = markov_cycle,
-    cycle_length = markov_cycle_length, 
-    type = "surv"
-  )
-  
-  res <- cbind(
-    pfs_surv,
-    os_surv - pfs_surv, 
-    1 - os_surv
-  )
-  
-  if ("terminal" %in% state_names) {
-    ## fix the "terminal" state
-    terminal <- c(0, diff(res[, 4]))
-    res[, 3:4] <- cbind(res[, 1:2], terminal, res[,3])
-  }
-  
-  res <- res * num_patients
-  colnames(res) <- state_names
-  res <- data.frame(res)
-  
-  structure(res, class = c("cycle_counts", class(res)))
-}
-
-#' Define Partitioned Survival
-#' 
-#' Define a partitioned survival model with progression-free
-#' survival and overall survival.
-#' 
-#' @param pfs Distribution of progression-free survival.
-#' @param os Distribution of overall survival.
-#' @param state_names character vector, optional, length 3. State
-#'   names for original state, progression and death
-#'   respectively.
-#' @param km_limit
-#'   
-#' @return
-#' @export
-#' 
-#' @examples
-define_part_surv <- function(pfs, os, state_names, km_limit) {
-  if (missing(state_names)) {
-    message("No named state -> generating names.")
-    state_names <- LETTERS[seq_len(3)]
-  } else {
-    if (length(state_names) != 3) {
-      stop("'state_names' should be of length 3.")
-    }
-  }
-  
 }