+++ R/tabular_input.R
@@ -265,25 +265,13 @@ create_model_list_from_tabular <- function(ref, df_env = globalenv()) {
   ## to accomodate partitioned survival models, we will allow for
   ##   the possibility that there is no transition matrix ...
   if (options()$heemod.verbose) message("*** Reading TM...")
-  tm_info <- NULL
-  tm_file <- ref$full_file[ref$data == "tm"]
-  if(length(tm_file) > 0 && file.exists(tm_file)){
-    tm_info <- read_file(ref$full_file[ref$data == "tm"])
-    tm_info <- parse_multi_spec(
-      tm_info,
-      group_vars = c("from", "to")
-    )
-  }
   
-  surv_info <- NULL
-  if("survivalInformation" %in% ref$data){
-    if(options()$heemod.verbose) 
-      message("** Reading survival model information..")
-    surv_info <- 
-      partitioned_survival_from_tabular(ref, df_env, names(state_info))
-  }
+  tm_info <- parse_multi_spec(
+    read_file(ref$full_file[ref$data == "tm"]),
+    group_vars = c("from", "to")
+  )
   
-  if (!is.null(tm_info) && length(pb <- setdiff(names(state_info), names(tm_info)))) {
+  if (length(pb <- setdiff(names(state_info), names(tm_info)))) {
     stop(sprintf(
       "Mismatching model names between TM file and state file: %s.",
       paste(pb, collapse = ", ")
@@ -291,14 +279,12 @@ create_model_list_from_tabular <- function(ref, df_env = globalenv()) {
   }
   
   tm_info <- tm_info[names(state_info)]
-  surv_info <- surv_info[names(state_info)]
   
   if (options()$heemod.verbose) message("*** Defining models...")
   models <- lapply(
     seq_along(state_info),
     function(i) {
       create_model_from_tabular(state_info[[i]], tm_info[[i]],
-                                surv_info[[i]],
                                 df_env = df_env)
     })  
   
@@ -669,9 +655,9 @@ create_options_from_tabular <- function(opt) {
 #' Calls \code{\link{create_states_from_tabular}} and
 #' \code{\link{create_matrix_from_tabular}}.
 #' 
-#' @param state_file A state tabular file (file path or 
+#' @param state_info A state tabular file (file path or 
 #'   parsed file).
-#' @param tm_file A transition matrix tabular file (file 
+#' @param tm_info A transition matrix tabular file (file 
 #'   path or parsed file).
 #' @param df_env An environment containing external data.
 #' 
@@ -681,32 +667,28 @@ create_options_from_tabular <- function(opt) {
 #' @keywords internal
 create_model_from_tabular <- function(state_info,
                                       tm_info,
-                                      surv_info,
                                       df_env = globalenv()) {
-  if(length(tm_info) == 0 & length(surv_info) == 0)
-    stop("either a transition matrix (tm_info) or a partitioned
-         survival model (surv_info) must be defined")
-  if(! inherits(state_info, "data.frame"))
-    stop("'state_info' must be a data frame.")
+  if(length(tm_info) == 0) {
+    stop("A transition matrix must be defined.")
+  }
   
-  if(!is.null(tm_info) && ! inherits(tm_info, "data.frame"))
-    stop("'tm_info' must be a data frame if defined.")
-  ##  if(!is.null(surv_info) && ! inherits(surv_info, "data.frame"))
-  ##    stop("'surv_info' must be a data frame if defined.")
+  if(! inherits(state_info, "data.frame")) {
+    stop("'state_info' must be a data frame.")
+  }
+  
+  if(!is.null(tm_info) && ! inherits(tm_info, "data.frame")) {
+    stop("'tm_info' must be a data frame.")
+  }
   
   if (options()$heemod.verbose) message("**** Defining state list...")
   states <- create_states_from_tabular(state_info,
                                        df_env = df_env)
   if (options()$heemod.verbose) message("**** Defining TM...")
   
-  TM <- "undefined"
-  if(!is.null(tm_info))
-    TM <- create_matrix_from_tabular(tm_info, get_state_names(states),
-                                     df_env = df_env)
+  TM <- create_matrix_from_tabular(tm_info, get_state_names(states),
+                                   df_env = df_env)
   
-  
-  define_strategy_(transition = TM, states = states, 
-                   partitioned_survival = surv_info)
+  define_strategy_(transition = TM, states = states)
 }
 
 #' Load Data From a Folder Into an Environment
@@ -755,206 +737,6 @@ create_df_from_tabular <- function(df_dir, df_envir) {
     assign(obj_names[i], this_val, envir = df_envir)
   }
   df_envir
-}
-
-#' Create Survival Analysis Curves from Tabular Data
-#' 
-#' @param base_dir base directory for the analysis.
-#' @param surv_dir subdirectory in which survival data and
-#'   fits are stored.
-#' @param data_files files containing survival data.
-#' @param fit_files Files containing fits.
-#' @param fit_names name(s) to save the fits to, if they are
-#'   being saved in a file
-#' @param time_col_name,censor_col_name,treatment_col_name 
-#'   columns to be used in the analysis
-#' @param fit_metric  Once of AIC, BIC, or m2LL
-#' @param best_only should only the best fit for each
-#'   condition be returned?
-#' @param dists the distributions to use to fit the survival
-#'   function
-#' @param use_envir An environment for the results to be
-#'   saved in.
-#'   
-#' @return A list with two elements:  \itemize{ 
-#'   \item{\code{best_models}, a list with the fits for each
-#'   data file passed in; and} \item{\code{envir}, an
-#'   environment containing the models so they can be
-#'   referenced to get probabilities.} }
-#' @details By default, the function fits with six different
-#'   distribution fucntions: exponential,  Weibull, 
-#'   lognormal, Gompertz, gamma, and generalized gamma.
-#'   
-#'   If data_files is NULL and fit files exists, then
-#'   fit_files should have the names of files in which
-#'   survival models are kept. If data_files is not NULL,
-#'   then survival models will be fit from the data in
-#'   data_files (using the \code{flexsurvreg} package), and
-#'   if fit_files is also not NULL, the survival models will
-#'   be saved in fit_files.
-#'   
-#'   If \code{best_only = TRUE}, then \code{best_models}
-#'   will be a vector of models, one for each group. If
-#'   \code{best_only = FALSE}, \code{best_models} will have
-#'   a matrix of models, with each row corresponding to a 
-#'   distribution and each column to a group.
-create_survival_from_tabular <- function(surv_dir,
-                                         time_col_name = "time",
-                                         censor_col_name = "status",
-                                         treatment_col_name = "treatment",
-                                         metric = "AIC",
-                                         best_only = TRUE,
-                                         dists = c("exp", "weibull", "lnorm", "gamma", 
-                                                   "gompertz", "gengamma"),
-                                         use_envir = NULL) {
-  
-  dists <- match.args(dists, several.ok = TRUE)
-  
-  if (length(data_files) == 0 & length(fit_files) == 0) {
-    stop("must specify at least one of data_files or fit_files")
-  }
-  
-  if (length(fit_files) > 0 & 
-      length(fit_files) != length(fit_names)) {
-    stop(paste("must specify the same number of fit_files and",
-               "fit_names if any fit_names are specified."))
-  }
-  
-  if (!is.null(base_dir)) {
-    full_path <- file.path(base_dir, surv_dir)
-  } else {
-    full_path <- surv_dir
-  }
-  
-  if (length(data_files) > 0) {
-    ## read data from files and fit models 
-    survdata <- lapply(data_files, function(this_file) {
-      ## if the "file name" begins with "list(", then
-      ##   it just needs to be evaluated into a list.
-      really_a_list_expr <- substring(this_file, 1, 5) == "list("
-      if(really_a_list_expr){
-        res <- eval(parse(text = this_file))
-      } else {
-        res <- read_file(file.path(full_path, this_file))
-        class(res) <- c("survdata", class(res))
-      }
-      res
-    })
-    ## now if we have a list from above, leave it alone; 
-    ##  otherwise fit the data
-    best_models <- lapply(survdata, function(this_data){
-      if(class(this_data)[1] == "list") {
-        this_data
-      } else {
-        surv_models <- fit_survival_models(
-          this_data,
-          dists = dists,
-          time_col_name = time_col_name, 
-          censor_col_name = censor_col_name, 
-          treatment_col_name = treatment_col_name,
-          covariate_col_names = NULL, 
-          fit_indiv_groups = TRUE
-        )
-        if (best_only) {
-          apply(surv_models, 2, f_get_best_surv_model, 
-                metric = fit_metric)
-        } else {
-          surv_models
-        }
-      }
-    })
-    ## best_dists <- lapply(best_models, dist_from_fits)
-    if (length(fit_files) > 0) {
-      ## eventually may want a different format for saving these
-      ## currently dump for a readable version, save for a complete version
-      for(i in seq(along = fit_names)){
-        assign(fit_names[i], best_models[[i]])
-        write_to_me <- paste(file.path(full_path, fit_files[i]), ".R", sep = "")
-        suppressMessages(
-          ## suppressMessages because we know we're going to get a warning
-          ##   that it may not be possible to completely deparse an object
-          ##   from flexsurv
-          ## dump(c(fit_names[i], "best_dists"), file = write_to_me)
-          dump(c(fit_names[i]), file = write_to_me)
-        )
-        save_to_me <- gsub(".R$", ".Rdata", write_to_me)
-        ## save(list = c(fit_names[i], "best_dists"), file = save_to_me)
-        save(list = c(fit_names[i]), file = save_to_me)
-      }
-    }
-    if(!is.null(use_envir))
-      for(i in seq(along = fit_names))
-        assign(fit_names[i], best_models[[i]], envir = use_envir)
-    
-  } else {   # that is, if we did not get data files - we'll load previous fits
-    for(this_fit_file in fit_files){
-      load(paste(file.path(full_path, this_fit_file), ".RData", sep = ""))
-      if(!is.null(use_envir))
-        load(paste(file.path(full_path, this_fit_file), ".RData", sep = ""), 
-             envir = use_envir)
-    }
-    best_models <- mget(fit_names)
-  }
-  
-  names(best_models) <- fit_names
-  res <- list(best_models, envir = use_envir)
-  res
-}
-
-#' Create a partitioned survival model object from tabular
-#' input
-#' 
-#' @param ref Name of the reference file.
-#' @param df_env an environment in which survival results
-#'   will be stored.
-#' @param model_names names of the models.
-#'   
-#' @return a list, with one element for each model.  Each
-#'   element of the list is also a list, with one element
-#'   for PFS and one for OS.
-create_part_surv_from_tabular <- function(surv_dir, df_env, model_names) {
-  si <- get_survival_input(read_file(surv_ref_full_file)) 
-  
-  ## load or fit survival models into the environment
-  surv_inputs <- survival_from_data(
-    location, #ref[ref$data == "survivalDataDirectory", "full_file"],
-    si$surv_data_dir,
-    data_files = si$surv_data_files,
-    fit_files = si$fit_files,
-    fit_names = si$fit_names,
-    fit_metric = si$fit_metric,
-    time_col_name = si$time_col_name,
-    censor_col_name = si$censor_col_name,
-    treatment_col_name = si$treatment_col_name,
-    dists = si$dists,
-    use_envir = df_env)
-  
-  if(! all(model_names %in% names(surv_inputs[[1]]$OS.fit))){
-    if(inherits(surv_inputs[[1]]$OS.fit[[1]],"flexsurvreg")){
-      stop(paste("some model names are not included in the survival models:\n",
-                 paste(model_names, collapse = ","),
-                 "vs.",
-                 paste(names(surv_inputs[[1]]$OS.fit), collapse = ","),
-                 "\n",
-                 "could treatment_col_name be wrong?"))
-    }
-    if(inherits(surv_inputs[[1]]$OS.fit[[1]], "list")){
-      stop(paste("some model names are not included in the survival models:\n",
-                 paste(model_names, collapse = ","),
-                 "vs.",
-                 paste(names(surv_inputs[[1]]$OS.fit), collapse = ","),
-                 "\n",
-                 "are the distribution lists properly specified?"))
-    }
-  }
-  
-  ## switch the survival information to be by model instead of by fit type
-  surv_package <- lapply(model_names, function(this_model) {
-    list(os = surv_inputs[[1]]$OS.fit[[this_model]],
-         pfs = surv_inputs[[1]]$PFS.fit[[this_model]])
-  })
-  names(surv_package) <- model_names
-  surv_package
 }
 
 #'Specify Inputs for Multiple Models From a Single File
@@ -1313,137 +1095,4 @@ save_graph <- function(plot, path, file_name) {
   )
   print(plot)
   grDevices::dev.off()
-}
-
-
-#' Make sure survival inputs have the correct formats
-#'
-#' @param surv_ref - data frame with survival data information
-#'
-#' @details For survival analysis, we need, for each condition
-#'   (frequently progression-free survival and overall survival)
-#'   at least two, and possibly three, elements represented in
-#'   the survival reference file.
-#'   \itemize{
-#'     \item{\code{surv_data_file}:  the name of the file 
-#'        with the survival data}
-#'     \item{\code{fit_file}: the name of a file with fits.  If
-#'     \code{surv_data_file} is not provided, the fit will
-#'     be read in from this file; if \code{surv_data_file}
-#'     is provided, the fit will be written into this file.}
-#'     \item{\code{fit_name}:  name to give the survival fit so
-#'     that it can be referred to to extract probabilities.}
-#'     } 
-#' At least one of \code{surv_data_file} and \code{fit_file}
-#'   must be provided.
-#' @return a list with elements \itemize{\item{fit_files},
-#'   \item{fit_names}, \item{surv_data_files}, \item{fit_metric}}.
-#'
-get_survival_input <- function(surv_ref) {
-  
-  if(! identical(names(surv_ref), c("data", "val")))
-    stop(
-      paste("surv_ref must have column names 'data' and 'val'",
-            paste("(actual_names", paste(names(surv_ref),collapse = ", "),
-                  ")"
-            )
-      )
-    )
-  surv_data_dir <- 
-    surv_ref[surv_ref$data == "survivalDataDirectory", "val"]
-  time_col_name <- "time"
-  censor_col_name <- "status"
-  treatment_col_name <- "treatment"
-  
-  dists <- c("exp", "weibull", "lnorm", "gamma", 
-             "gompertz", "gengamma")
-  
-  time_col_index <- grep("time_col_name", surv_ref$data)
-  censor_col_index <- grep("censor_col_name", surv_ref$data)
-  treatment_col_index <- grep("treatment_col_name", surv_ref$data)
-  dists_index <- grep("dists", surv_ref$data)
-  
-  if(length(time_col_index) > 1)
-    stop("must have at most one time_col_name; default = 'time'")
-  if(length(censor_col_index) > 1)
-    stop("must have at most one censor_col_name; default = 'status'")
-  if(length(treatment_col_index) > 1)
-    stop("must have at most one treatment_col_name; default = 'treatment'")
-  if(length(dists_index) > 1)
-    stop(paste('must have at most one line specificying distributions;\n',
-               'default = c("exp", "weibull", "lnorm", "gamma", 
-                           "gompertz", "gengamma")'))
-  
-  if(length(time_col_index) == 1)
-    time_col_name <- surv_ref[time_col_index, "val"]
-  if(length(censor_col_index) == 1)
-    censor_col_name <- surv_ref[censor_col_index, "val"]
-  if(length(treatment_col_index) == 1)
-    treatment_col_name <- surv_ref[treatment_col_index, "val"]
-  if(length(dists_index) ==1)
-    dists <- surv_ref[dists_index, "val"]
-  
-  surv_data_indices <- grep("surv_data_file", surv_ref$data)
-  fit_file_indices <- grep("fit_file", surv_ref$data)
-  fit_name_indices <- grep("fit_name", surv_ref$data)
-  
-  surv_data_names <- surv_ref[surv_data_indices, "data"]
-  fit_file_names <- surv_ref[fit_file_indices, "data"]
-  fit_name_names <- surv_ref[fit_name_indices, "data"]
-  
-  surv_data_files = as.vector(surv_ref[surv_data_indices, "val"])
-  fit_files = as.vector(surv_ref[fit_file_indices, "val"])
-  fit_names = as.vector(surv_ref[fit_name_indices, "val"])
-  
-  fit_metric = surv_ref[surv_ref$data == "fit_metric", "val"]
-  
-  lff <- length(fit_file_names)
-  lsdn <- length(surv_data_names)
-  
-  if(lff == 0 & lsdn == 0)
-    stop(paste("survival reference file must define at least one of",
-               "fit_file and surv_data_file for survival analysis"))
-  if(lff != 0 & lsdn != 0 & (lff != lsdn))
-    stop(paste("when both fit_file and surv_data_name elements are",
-               "specified,\nshould have the same number of elements"))
-  
-  if(length(fit_names) != max(lff, lsdn))
-    stop(paste("ref file should have the same number of fit_name", 
-               "elements as fit_file or surv_data_file elements"))
-  
-  suffixes <- cbind(surv_data_names, fit_name_names, fit_file_names)
-  
-  suffixes <- gsub("fit_file", "", suffixes)
-  suffixes <- gsub("fit_name", "", suffixes)
-  suffixes <- gsub("surv_data_file", "", suffixes)
-  
-  
-  for (i in 1:ncol(suffixes))
-    suffixes[, i] <- sort(suffixes[, i])
-  num_suffixes <- apply(suffixes, 1, function(x) {
-    length(unique(x))
-  })
-  if (any(num_suffixes != 1)){
-    print(table(suffixes))
-    stop("suffixes do not match")
-  }
-  
-  fit_files <- fit_files[order(fit_file_names)]
-  fit_names <- fit_names[order(fit_name_names)]
-  surv_data_files <- surv_data_files[order(surv_data_names)]
-  
-  return(
-    list(
-      surv_data_dir = surv_data_dir, 
-      fit_files = fit_files,
-      fit_names = fit_names,
-      surv_data_files = surv_data_files,
-      fit_metric = fit_metric,
-      time_col_name = time_col_name,
-      censor_col_name = censor_col_name,
-      treatment_col_name = treatment_col_name,
-      dists = dists
-    )
-  )
-  
 }